backend/requirements.txt
openai-whisper
kokoro
numpy
soundfile
langchain-ollama
langchain-core
langchain-mongodb
pynput